Fixed 5

+++ zinoviev.alexander/S1/main.cpp
@@ -8,13 +8,13 @@ using namespace zinoviev;
 
 int main()
 {
-  BiList<std::pair<std::string, BiList<int>>> outer;
+  BiList<std::pair<std::string, BiList<unsigned long long>>> outer;
   std::string name;
-  long long x;
+  unsigned long long x;
 
   while (std::cin >> name)
   {
-    BiList<int> inner;
+    BiList<unsigned long long> inner;
 
     while (std::cin >> x)
     {
@@ -23,7 +23,7 @@ int main()
         std::cerr << "Error: number out of range\n";
         return 1;
       }
-      inner.push_back(static_cast<int>(x));
+      inner.push_back(static_cast<unsigned long long>(x));
     }
 
     outer.push_back(std::make_pair(name, std::move(inner)));
@@ -40,7 +40,7 @@ int main()
     return 0;
   }
 
-  CBIter<std::pair<std::string, BiList<int>>> c_iter_outer = outer.cbegin();
+  CBIter<std::pair<std::string, BiList<unsigned long long>>> c_iter_outer = outer.cbegin();
 
   std::cout << c_iter_outer->first;
   ++c_iter_outer;
@@ -52,9 +52,9 @@ int main()
   }
   std::cout << "\n";
 
-  BIter<std::pair<std::string, BiList<int>>> iter_outer = outer.begin();
+  BIter<std::pair<std::string, BiList<unsigned long long>>> iter_outer = outer.begin();
 
-  BiList<std::pair<BIter<int>, BIter<int>>> list_iter_inner;
+  BiList<std::pair<BIter<unsigned long long>, BIter<unsigned long long>>> list_iter_inner;
   while (iter_outer != outer.end())
   {
     if (iter_outer->second.size() > 0)
@@ -70,10 +70,10 @@ int main()
     return 1;
   }
 
-  BIter<std::pair<BIter<int>, BIter<int>>> iter_l_i_inner = list_iter_inner.begin();
+  BIter<std::pair<BIter<unsigned long long>, BIter<unsigned long long>>> iter_l_i_inner = list_iter_inner.begin();
 
-  BiList<long long> list_sum;
-  long long sum = 0;
+  BiList<unsigned long long> list_sum;
+  unsigned long long sum = 0;
 
   while (list_iter_inner.size())
   {
@@ -84,11 +84,11 @@ int main()
 
     ++(iter_l_i_inner->first);
 
-    for (BIter<std::pair<BIter<int>, BIter<int>>> i = ++(list_iter_inner.begin()); i != list_iter_inner.end(); )
+    for (BIter<std::pair<BIter<unsigned long long>, BIter<unsigned long long>>> i = ++(list_iter_inner.begin()); i != list_iter_inner.end(); )
     {
       std::cout << " " << *(i->first);
 
-      if (sum > std::numeric_limits<long long>::max() - *(i->first))
+      if (sum > std::numeric_limits<unsigned long long>::max() - *(i->first))
       {
         std::cerr << "Error: overflow\n";
         return 1;
@@ -111,7 +111,7 @@ int main()
 
     if (iter_l_i_inner->first == iter_l_i_inner->second)
     {
-      BIter<std::pair<BIter<int>, BIter<int>>> tmp = iter_l_i_inner;
+      BIter<std::pair<BIter<unsigned long long>, BIter<unsigned long long>>> tmp = iter_l_i_inner;
       ++iter_l_i_inner;
       list_iter_inner.erase(tmp);
     }
@@ -119,7 +119,7 @@ int main()
     list_sum.push_back(sum);
   }
 
-  CBIter<long long> iter_sum = list_sum.cbegin();
+  CBIter<unsigned long long> iter_sum = list_sum.cbegin();
 
   std::cout << *iter_sum;
   ++iter_sum;
